@@ -1,33 +1,40 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
-        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch": "npm run development -- --watch",
-        "watch-poll": "npm run watch -- --watch-poll",
-        "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --disable-host-check --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "dev": "mix",
+        "development": "mix",
+        "watch": "mix watch",
+        "watch-poll": "mix watch -- --watch-options-poll=1000",
+        "hot": "mix watch --hot",
+        "prod": "mix --production",
+        "production": "mix --production"
     },
     "devDependencies": {
         "@babel/core": "^7.13.15",
         "@babel/preset-env": "^7.13.15",
+        "@vue/compiler-sfc": "^3.0.11",
         "axios": "^0.21.1",
         "babel-loader": "^8.2.2",
         "bootstrap": "^4.6.0",
         "cross-env": "^7.0",
         "jquery": "^3.6.0",
-        "laravel-mix": "^5.0.1",
+        "laravel-mix": "^6.0.19",
         "lodash": "^4.17.21",
+        "node-sass": "^4.14.1",
         "popper.js": "^1.12",
         "resolve-url-loader": "^3.1.0",
         "sass": "^1.32.8",
         "sass-loader": "^8.0.0",
+        "vue": "^3.0.11",
+        "vue-loader": "^16.2.0",
         "vue-template-compiler": "^2.6.12",
-        "webpack": "^5.31.0"
+        "webpack": "^5.36.2"
     },
     "dependencies": {
         "@fortawesome/fontawesome-free": "^5.15.3",
+        "@inertiajs/inertia": "^0.8.7",
+        "@inertiajs/inertia-vue3": "^0.3.14",
+        "@inertiajs/progress": "^0.2.4",
         "laravel-bootstrap-table": "0.0.5"
     }
 }
